+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.5.0] - 2020-06-08
+
+### Changed
+- `Unity.Platforms.Windows` assembly renamed to `Unity.Build.Windows.DotsRuntime`.
+- `Unity.Platforms.Windows.Build` assembly renamed to `Unity.Build.Windows.Classic`.
+
+## [0.4.1] - 2020-05-27
+
+### Changed
+- Update platforms packages to 0.4.1-preview
+- Updated minimum Unity version to 2020.1.
+
 ## [0.4.0] - 2020-05-05
 
 ### Changed